Cmmn: create default event dispatcher on engine initialisation
